requirements.txt
flask==2.0.3
flask-smorest==0.36.0
marshmallow-sqlalchemy==0.27.0
Flask-SQLAlchemy==2.5.1
flask-marshmallow==0.14.0
python-slugify==6.0.1
pytest==7.0.1
mysqlclient==2.1.0